Added the grocery list view so you can see items in a grocery list
This adds the click handler to be able to open the grocery list, and the
classes to view grocery list. This is entirely mocked and there is no
real data yet. There is also a click handler for when an item is
clicked, but this is unsued. It was accidently made. It was supposed to
be for the main controller. It was left in because it will be used soon
enough.