Added a create grocery list dialog
The main activity now uses the floating action button to display a
dialog to create a new list. The create is just mocked out in memory.
The main activity is currently used for the grocery list, but it will be
changed to its own activity. This means that today, you can create a
grocery list inside of the grocery list, but once again, this will be
removed.