Implemented loading the grocery list from firestore
This just adds support for loading the list from firestore, not adding
new items or changing existing items. On a side note, I don't now if the
current format is really going to work well. It might cause a bit of
redudant saving and network usage. To fix we might want to break out the
list items from the list itself. I don't know if I will ever get to it
though considering the time till this has to be done.